feat: move to pkg
