+++ apis/address/address.go
@@ -1,0 +1,156 @@
+package address
+
+import (
+	"fmt"
+
+	cfd "github.com/cryptogarageinc/cfd-go"
+	config "github.com/cryptogarageinc/cfd-go/config"
+	types "github.com/cryptogarageinc/cfd-go/types"
+)
+
+// -------------------------------------
+// API
+// -------------------------------------
+
+type AddressApi interface {
+	ParseAddress(addressString string) (address *types.Address, err error)
+	CreateByPubkey(pubkey *types.Pubkey, addressType types.AddressType) (address *types.Address, err error)
+	CreateByScript(redeemScript *types.Script, addressType types.AddressType) (address *types.Address, err error)
+	CreateMultisigAddress(pubkeys *[]types.Pubkey, requireNum uint32, addressType types.AddressType) (address *types.Address, redeemScript *types.Script, err error)
+	GetPeginAddressByPubkey(addressType types.AddressType, fedpegScript, pubkey string) (peginAddress *types.Address, claimScript *types.Script, err error)
+	GetPegoutAddress(addressType types.AddressType, descriptorOrXpub string, bip32Counter uint32) (pegoutAddress *types.Address, baseDescriptor *string, err error)
+}
+
+func NewAddressApi() AddressApi {
+	return &AddressUtil{}
+}
+
+// -------------------------------------
+// AddressUtil
+// -------------------------------------
+
+// AddressUtil ...
+type AddressUtil struct {
+	Network *types.NetworkType
+}
+
+// ParseAddress ...
+func (u *AddressUtil) ParseAddress(addressString string) (address *types.Address, err error) {
+	data, err := cfd.CfdGoGetAddressInfo(addressString)
+	if err != nil {
+		return nil, err
+	}
+	return &types.Address{
+		Address: addressString,
+		Network: types.NewNetworkType(data.NetworkType),
+		Type:    types.NewAddressTypeByHashType(data.HashType),
+	}, nil
+}
+
+// CreateByPubkey ...
+func (u *AddressUtil) CreateByPubkey(pubkey *types.Pubkey, addressType types.AddressType) (address *types.Address, err error) {
+	if err = u.validConfig(); err != nil {
+		return nil, err
+	}
+	addr, _, _, err := cfd.CfdGoCreateAddress(addressType.ToHashType().ToCfdValue(), pubkey.Hex, "", u.Network.ToCfdValue())
+	if err != nil {
+		return nil, err
+	}
+	return &types.Address{
+		Address: addr,
+		Network: *u.Network,
+		Type:    addressType,
+	}, nil
+}
+
+// CreateByScript ...
+func (u *AddressUtil) CreateByScript(redeemScript *types.Script, addressType types.AddressType) (address *types.Address, err error) {
+	if err = u.validConfig(); err != nil {
+		return nil, err
+	}
+	addr, _, _, err := cfd.CfdGoCreateAddress(addressType.ToHashType().ToCfdValue(), "", redeemScript.ToHex(), u.Network.ToCfdValue())
+	if err != nil {
+		return nil, err
+	}
+	return &types.Address{
+		Address: addr,
+		Network: *u.Network,
+		Type:    addressType,
+	}, nil
+}
+
+// CreateMultisigAddress ...
+func (u *AddressUtil) CreateMultisigAddress(pubkeys *[]types.Pubkey, requireNum uint32, addressType types.AddressType) (address *types.Address, redeemScript *types.Script, err error) {
+	if err = u.validConfig(); err != nil {
+		return nil, nil, err
+	}
+	pubkeyList := make([]string, len(*pubkeys))
+	for i := 0; i < len(*pubkeys); i++ {
+		pubkeyList[i] = (*pubkeys)[i].Hex
+	}
+	addr, script, witnessScript, err := cfd.CfdGoCreateMultisigScript(u.Network.ToCfdValue(), addressType.ToHashType().ToCfdValue(), pubkeyList, requireNum)
+	if err != nil {
+		return nil, nil, err
+	}
+	if addressType == types.P2shAddress {
+		redeemScript = types.NewScriptFromHexIgnoreError(script)
+	} else {
+		redeemScript = types.NewScriptFromHexIgnoreError(witnessScript)
+	}
+	return &types.Address{
+		Address: addr,
+		Network: *u.Network,
+		Type:    addressType,
+	}, redeemScript, nil
+}
+
+// GetPeginAddressByPubkey ...
+func (u *AddressUtil) GetPeginAddressByPubkey(addressType types.AddressType, fedpegScript, pubkey string) (peginAddress *types.Address, claimScript *types.Script, err error) {
+	if err = u.validConfig(); err != nil {
+		return nil, nil, err
+	}
+
+	addr, script, _, err := cfd.GetPeginAddress(u.Network.ToBitcoinType().ToCfdValue(), fedpegScript, addressType.ToCfdValue(), pubkey, "")
+	if err != nil {
+		return nil, nil, err
+	}
+	peginAddress = &types.Address{
+		Address: addr,
+		Network: u.Network.ToBitcoinType(),
+		Type:    addressType,
+	}
+	claimScript = types.NewScriptFromHexIgnoreError(script)
+	return peginAddress, claimScript, nil
+}
+
+// GetPegoutAddress ...
+func (u *AddressUtil) GetPegoutAddress(addressType types.AddressType, descriptorOrXpub string, bip32Counter uint32) (pegoutAddress *types.Address, baseDescriptor *string, err error) {
+	if err = u.validConfig(); err != nil {
+		return nil, nil, err
+	}
+
+	addr, desc, err := cfd.GetPegoutAddress(u.Network.ToBitcoinType().ToCfdValue(), u.Network.ToCfdValue(), descriptorOrXpub, bip32Counter, addressType.ToCfdValue())
+	if err != nil {
+		return nil, nil, err
+	}
+	pegoutAddress = &types.Address{
+		Address: addr,
+		Network: u.Network.ToBitcoinType(),
+		Type:    addressType,
+	}
+	baseDescriptor = &desc
+	return pegoutAddress, baseDescriptor, nil
+}
+
+// validConfig ...
+func (u *AddressUtil) validConfig() error {
+	if u.Network == nil {
+		cfdConfig := config.GetCurrentCfdConfig()
+		if !cfdConfig.Network.Valid() {
+			return fmt.Errorf("CFD Error: NetworkType not set")
+		}
+		netType := cfdConfig.Network
+		u.Network = &netType
+	}
+	return nil
+}
