+++ apis/script/script.go
@@ -1,0 +1,5 @@
+package script
+
+// Script This struct holds a script.
+type ScriptUtil struct {
+}
