+++ apis/transaction/elements_tx_test.go
@@ -1,27 +1,43 @@
-package cfdgo
+package transaction
 
 import (
 	"fmt"
+	"runtime"
+	"strings"
 	"testing"
 
+	cfd "github.com/cryptogarageinc/cfd-go"
+	"github.com/cryptogarageinc/cfd-go/apis/address"
+	"github.com/cryptogarageinc/cfd-go/apis/key"
+	"github.com/cryptogarageinc/cfd-go/config"
+	"github.com/cryptogarageinc/cfd-go/types"
 	"github.com/stretchr/testify/assert"
 )
 
+// GetFuncName
+func GetFuncName() string {
+	pc, _, _, _ := runtime.Caller(1)
+	funcName := runtime.FuncForPC(pc).Name()
+	index := strings.LastIndex(funcName, ".")
+	return funcName[index+1:]
+}
+
 func TestCreateClaimPeginTx(t *testing.T) {
-	network := ElementsRegtest
+	network := types.ElementsRegtest
 	genesisBlockHash := "0f9188f13cb7b2c71f2a335e3a4fc328bf5beb436012afca590b1a11466e2206"
 	asset := "5ac9f65c0efcc4775e0baec4ec03abdde22473cd3cf33c0419ca290e0751b225"
 
 	// fedpeg script
 	fedpegScript := "522103aab896d53a8e7d6433137bbba940f9c521e085dd07e60994579b64a6d992cf79210291b7d0b1b692f8f524516ed950872e5da10fb1b808b5a526dedc6fed1cf29807210386aa9372fbab374593466bc5451dc59954e90787f08060964d95c87ef34ca5bb53ae"
-	privkey, err := NewPrivkeyFromWif("cUfipPioYnHU61pfYTH9uuNoswRXx8rtzXhJZrsPeVV1LRFdTxvp")
+	privkey, err := key.NewPrivkeyFromWif("cUfipPioYnHU61pfYTH9uuNoswRXx8rtzXhJZrsPeVV1LRFdTxvp")
 	assert.NoError(t, err)
-	pubkey, err := privkey.GetPubkey()
+	keyUtil := &key.PrivkeyUtil{}
+	pubkey, err := keyUtil.GetPubkey(privkey)
 	assert.NoError(t, err)
 
 	// create pegin address
-	addrUtil := AddressUtil{Network: &network}
-	peginAddr, claimScript, err := addrUtil.GetPeginAddressByPubkey(P2shP2wshAddress, fedpegScript, pubkey.Hex)
+	addrUtil := address.AddressUtil{Network: &network}
+	peginAddr, claimScript, err := addrUtil.GetPeginAddressByPubkey(types.P2shP2wshAddress, fedpegScript, pubkey.Hex)
 	assert.NoError(t, err)
 	assert.Equal(t, "2MvmzAFKZ5xh44vyb7qY7NB2AoDuS55rVFW", peginAddr.Address)
 	assert.Equal(t, "0014e794713e386d83f32baa0e9d03e47c0839dc57a8", claimScript.ToHex())
@@ -32,17 +48,17 @@ func TestCreateClaimPeginTx(t *testing.T) {
 	peginAmount := amount - feeAmount
 	btcNetwork := network.ToBitcoinType()
 	btcTxUtil := TransactionUtil{Network: &btcNetwork}
-	utxoOutPoint := OutPoint{
+	utxoOutPoint := types.OutPoint{
 		Txid: "ea9d5a9e974af1d167305aa6ee598706d63274e8a40f4f33af97db37a7adde4c",
 		Vout: 0,
 	}
-	btcInputs := []InputTxIn{
+	btcInputs := []types.InputTxIn{
 		{
 			OutPoint: utxoOutPoint,
-			Sequence: SequenceLockTimeFinal,
+			Sequence: types.SequenceLockTimeFinal,
 		},
 	}
-	btcOutputs := []InputTxOut{
+	btcOutputs := []types.InputTxOut{
 		{
 			Amount:  peginAmount,
 			Address: peginAddr.Address,
@@ -51,7 +67,7 @@ func TestCreateClaimPeginTx(t *testing.T) {
 	btcTx, err := btcTxUtil.Create(uint32(2), uint32(0), &btcInputs, &btcOutputs)
 	assert.NoError(t, err)
 	// add sign
-	utxos := []UtxoData{
+	utxos := []types.UtxoData{
 		{
 			Txid:       utxoOutPoint.Txid,
 			Vout:       utxoOutPoint.Vout,
@@ -59,27 +75,27 @@ func TestCreateClaimPeginTx(t *testing.T) {
 			Descriptor: "wpkh(02fd54c734e48c544c3c3ad1aab0607f896eb95e23e7058b174a580826a7940ad8)",
 		},
 	}
-	utxoPrivkey, err := NewPrivkeyFromWif("cNYKHjNc33ZyNMcDck59yWm1CYohgPhr2DYyCtmWNkL6sqb5L1rH")
+	utxoPrivkey, err := key.NewPrivkeyFromWif("cNYKHjNc33ZyNMcDck59yWm1CYohgPhr2DYyCtmWNkL6sqb5L1rH")
 	assert.NoError(t, err)
-	err = btcTx.SignWithPrivkey(&utxoOutPoint, utxoPrivkey, SigHashTypeAll, &utxos)
+	err = btcTxUtil.SignWithPrivkey(btcTx, &utxoOutPoint, utxoPrivkey, types.SigHashTypeAll, &utxos)
 	assert.NoError(t, err)
 	assert.Equal(t, "020000000001014cdeada737db97af334f0fa4e87432d6068759eea65a3067d1f14a979e5a9dea0000000000ffffffff010cdff5050000000017a91426b9ba9cf5d822b70cf490ad0394566f9db20c63870247304402200b3ca71e82551a333fe5c8ce9a8f8454eb8f08aa194180e5a87c79ccf2e46212022065c1f2a363ebcb155a80e234258394140d08f6ab807581953bb21a58f2d229a6012102fd54c734e48c544c3c3ad1aab0607f896eb95e23e7058b174a580826a7940ad800000000", btcTx.Hex)
 
-	assert.Equal(t, "12708508f0baf8691a3d7e22fd19afbf9bd8bf0d358e3310838bcc7916539c7b", btcTx.GetTxid())
+	assert.Equal(t, "12708508f0baf8691a3d7e22fd19afbf9bd8bf0d358e3310838bcc7916539c7b", btcTxUtil.GetTxid(btcTx))
 
 	peginIndex := uint32(0)
 	txoutProof := "00000020fe3b574c1ce6d5cb68fc518e86f7976e599fafc0a2e5754aace7ca16d97a7c78ef9325b8d4f0a4921e060fc5e71435f46a18fa339688142cd4b028c8488c9f8dd1495b5dffff7f200200000002000000024a180a6822abffc3b1080c49016899c6dac25083936df14af12f58db11958ef27926299350fdc2f4d0da1d4f0fbbd3789d29f9dc016358ae42463c0cebf393f30105"
 
 	// create pegin tx
-	txUtil := ElementsTransactionUtil{Network: &network}
-	peginOutPoint := OutPoint{
-		Txid: btcTx.GetTxid(),
+	txUtil := ConfidentialTxUtil{Network: &network}
+	peginOutPoint := types.OutPoint{
+		Txid: btcTxUtil.GetTxid(btcTx),
 		Vout: peginIndex,
 	}
-	inputs := []InputConfidentialTxIn{
+	inputs := []types.InputConfidentialTxIn{
 		{
 			OutPoint: peginOutPoint,
-			PeginInput: &InputPeginData{
+			PeginInput: &types.InputPeginData{
 				BitcoinTransaction:      btcTx.Hex,
 				BitcoinGenesisBlockHash: genesisBlockHash,
 				BitcoinAssetId:          asset,
@@ -88,7 +104,7 @@ func TestCreateClaimPeginTx(t *testing.T) {
 			},
 		},
 	}
-	outputs := []InputConfidentialTxOut{
+	outputs := []types.InputConfidentialTxOut{
 		{
 			Amount:  99998500,
 			Address: "el1qqtl9a3n6878ex25u0wv8u5qlzpfkycc0cftk65t52pkauk55jqka0fajk8d80lafn4t9kqxe77cu9ez2dyr6sq54lwy009uex",
@@ -105,24 +121,24 @@ func TestCreateClaimPeginTx(t *testing.T) {
 	assert.Equal(t, "0200000001017b9c531679cc8b8310338e350dbfd89bbfaf19fd227e3d1a69f8baf0088570120000004000ffffffff020125b251070e29ca19043cf33ccd7324e2ddab03ecc4ae0b5e77c4fc0e5cf6c95a010000000005f5db2402fe5ec67a3f8f932a9c7b987e501f105362630fc2576d5174506dde5a94902dd7160014a7b2b1da77ffa99d565b00d9f7b1c2e44a6907a80125b251070e29ca19043cf33ccd7324e2ddab03ecc4ae0b5e77c4fc0e5cf6c95a0100000000000003e800000000000000000006080cdff505000000002025b251070e29ca19043cf33ccd7324e2ddab03ecc4ae0b5e77c4fc0e5cf6c95a2006226e46111a0b59caaf126043eb5bbf28c34f3a5e332a1fc7b2b73cf188910f160014e794713e386d83f32baa0e9d03e47c0839dc57a8c0020000000001014cdeada737db97af334f0fa4e87432d6068759eea65a3067d1f14a979e5a9dea0000000000ffffffff010cdff5050000000017a91426b9ba9cf5d822b70cf490ad0394566f9db20c63870247304402200b3ca71e82551a333fe5c8ce9a8f8454eb8f08aa194180e5a87c79ccf2e46212022065c1f2a363ebcb155a80e234258394140d08f6ab807581953bb21a58f2d229a6012102fd54c734e48c544c3c3ad1aab0607f896eb95e23e7058b174a580826a7940ad8000000009700000020fe3b574c1ce6d5cb68fc518e86f7976e599fafc0a2e5754aace7ca16d97a7c78ef9325b8d4f0a4921e060fc5e71435f46a18fa339688142cd4b028c8488c9f8dd1495b5dffff7f200200000002000000024a180a6822abffc3b1080c49016899c6dac25083936df14af12f58db11958ef27926299350fdc2f4d0da1d4f0fbbd3789d29f9dc016358ae42463c0cebf393f3010500000000", tx.Hex)
 
 	// blind
-	blindTxInList := []BlindInputData{
+	blindTxInList := []types.BlindInputData{
 		{
 			OutPoint: peginOutPoint,
 			Asset:    asset,
 			Amount:   peginAmount,
 		},
 	}
-	option := NewBlindTxOption()
+	option := types.NewBlindTxOption()
 	option.AppendDummyOutput = true
-	err = tx.Blind(blindTxInList, nil, &option)
+	err = txUtil.Blind(tx, blindTxInList, nil, &option)
 	assert.NoError(t, err)
-	_, inList, outList, err := tx.GetAll(false)
+	_, inList, outList, err := txUtil.GetAll(tx, false)
 	assert.NoError(t, err)
 	assert.Equal(t, 1, len(inList))
 	assert.Equal(t, 3, len(outList))
 
 	// sign
-	peginUtxos := []UtxoData{
+	peginUtxos := []types.UtxoData{
 		{
 			Txid:       peginOutPoint.Txid,
 			Vout:       peginOutPoint.Vout,
@@ -130,18 +146,22 @@ func TestCreateClaimPeginTx(t *testing.T) {
 			Descriptor: "wpkh(" + pubkey.Hex + ")",
 		},
 	}
-	sighash, err := tx.GetSighash(&peginOutPoint, SigHashTypeAll, &peginUtxos)
+	sighash, err := txUtil.GetSighash(tx, &peginOutPoint, types.SigHashTypeAll, &peginUtxos)
 	assert.NoError(t, err)
-	descriptor := NewDescriptorFromString(peginUtxos[0].Descriptor, network.ToCfdValue())
-	signature, err := privkey.CreateEcSignature(sighash, &SigHashTypeAll)
+	desc := types.Descriptor{
+		OutputDescriptor: peginUtxos[0].Descriptor,
+		Network:          &network,
+	}
+	privkeyUtil := &key.PrivkeyUtil{}
+	signature, err := privkeyUtil.CreateEcSignature(privkey, sighash, &types.SigHashTypeAll)
 	assert.NoError(t, err)
 
 	// add sign
-	err = tx.AddPubkeySignByDescriptor(&peginOutPoint, descriptor, signature.ToHex())
+	err = txUtil.AddPubkeySignByDescriptor(tx, &peginOutPoint, &desc, signature.ToHex())
 	assert.NoError(t, err)
 
 	// verify
-	isVerify, reason, err := tx.VerifySign(&peginOutPoint, peginAmount, &peginUtxos)
+	isVerify, reason, err := txUtil.VerifySign(tx, &peginOutPoint, peginAmount, &peginUtxos)
 	assert.NoError(t, err)
 	assert.True(t, isVerify)
 	assert.Equal(t, "", reason)
@@ -150,46 +170,46 @@ func TestCreateClaimPeginTx(t *testing.T) {
 }
 
 func TestCreatePegoutTx(t *testing.T) {
-	network := LiquidV1
+	network := types.LiquidV1
 	genesisBlockHash := "0f9188f13cb7b2c71f2a335e3a4fc328bf5beb436012afca590b1a11466e2206"
 	asset := "5ac9f65c0efcc4775e0baec4ec03abdde22473cd3cf33c0419ca290e0751b225"
 
 	// mainchain address descriptor
 	mainchainXpubkey := "xpub6A53gzNdDBQYCtFFpZT7kUpoBGpzWigaukrdF9xoUZt7cYMD2qCAHVLsstNoQEDMFJWdX78KvT6yxpC76aGCN5mENVdWtFGcWZoKdtLq5jW"
-	mainchainPubkey, err := CfdGoGetPubkeyFromExtkey(mainchainXpubkey, int(KCfdNetworkMainnet))
+	mainchainPubkey, err := cfd.CfdGoGetPubkeyFromExtkey(mainchainXpubkey, int(cfd.KCfdNetworkMainnet))
 	assert.NoError(t, err)
-	negateMainchainPubkey, err := CfdGoNegatePubkey(mainchainPubkey)
+	negateMainchainPubkey, err := cfd.CfdGoNegatePubkey(mainchainPubkey)
 	assert.NoError(t, err)
 	mainchainOutputDescriptor := "pkh(" + mainchainXpubkey + "/0/*)"
 	bip32Counter := uint32(0)
 
 	onlinePrivkey := "L52AgshDAE14NHJuovwAw8hyrTNK4YQjuiPC9EES4sfM7oBPzU4o"
-	onlinePubkey, err := CfdGoGetPubkeyFromPrivkey("", onlinePrivkey, true)
+	onlinePubkey, err := cfd.CfdGoGetPubkeyFromPrivkey("", onlinePrivkey, true)
 	assert.NoError(t, err)
 	// whitelist
 	pakEntry := negateMainchainPubkey + onlinePubkey
 	whitelist := pakEntry
 
 	// pegout address
-	addrUtil := AddressUtil{Network: &network}
-	pegoutAddr, baseDescriptor, err := addrUtil.GetPegoutAddress(P2pkhAddress, mainchainOutputDescriptor, bip32Counter)
+	addrUtil := address.AddressUtil{Network: &network}
+	pegoutAddr, baseDescriptor, err := addrUtil.GetPegoutAddress(types.P2pkhAddress, mainchainOutputDescriptor, bip32Counter)
 	assert.NoError(t, err)
 	assert.Equal(t, "1NrcpiZmCxjC7KVKAYT22SzVhhcXtp5o4v", pegoutAddr.Address)
 	assert.Equal(t, "pkh("+mainchainXpubkey+")", *baseDescriptor)
 
 	// create pegout tx
-	txUtil := ElementsTransactionUtil{Network: &network}
+	txUtil := ConfidentialTxUtil{Network: &network}
 	pegoutAddrList := []string{}
-	inputs := []InputConfidentialTxIn{
+	inputs := []types.InputConfidentialTxIn{
 		{
-			OutPoint: OutPoint{
+			OutPoint: types.OutPoint{
 				Txid: "4aa201f333e80b8f62ba5b593edb47b4730212e2917b21279f389ba1c14588a3",
 				Vout: 0,
 			},
 			Sequence: 4294967293,
 		},
 	}
-	outputs := []InputConfidentialTxOut{
+	outputs := []types.InputConfidentialTxOut{
 		{
 			Amount:  209998999992700,
 			Address: "XBMr6srTXmWuHifFd8gs54xYfiCBsvrksA",
@@ -198,7 +218,7 @@ func TestCreatePegoutTx(t *testing.T) {
 		{
 			Amount: 1000000000,
 			Asset:  asset,
-			PegoutInput: &InputPegoutData{
+			PegoutInput: &types.InputPegoutData{
 				OnlineKey:               onlinePrivkey,
 				BitcoinGenesisBlockHash: genesisBlockHash,
 				BitcoinOutputDescriptor: mainchainOutputDescriptor,
@@ -218,7 +238,7 @@ func TestCreatePegoutTx(t *testing.T) {
 	assert.Equal(t, 1, len(pegoutAddrList))
 	assert.Equal(t, "1NrcpiZmCxjC7KVKAYT22SzVhhcXtp5o4v", pegoutAddrList[0])
 
-	pegoutAddress, hasPegout, err := tx.GetPegoutAddress(uint32(1))
+	pegoutAddress, hasPegout, err := txUtil.GetPegoutAddress(tx, uint32(1))
 	assert.NoError(t, err)
 	assert.True(t, hasPegout)
 	assert.Equal(t, "1NrcpiZmCxjC7KVKAYT22SzVhhcXtp5o4v", pegoutAddress.Address)
@@ -227,22 +247,23 @@ func TestCreatePegoutTx(t *testing.T) {
 }
 
 func TestCreateClaimPeginTxByCfdConf(t *testing.T) {
-	SetCfdConfig(CfdConfig{
-		Network:                 ElementsRegtest,
+	config.SetCfdConfig(config.CfdConfig{
+		Network:                 types.ElementsRegtest,
 		BitcoinGenesisBlockHash: "0f9188f13cb7b2c71f2a335e3a4fc328bf5beb436012afca590b1a11466e2206",
 		BitcoinAssetId:          "5ac9f65c0efcc4775e0baec4ec03abdde22473cd3cf33c0419ca290e0751b225",
 	})
 
 	// fedpeg script
 	fedpegScript := "522103aab896d53a8e7d6433137bbba940f9c521e085dd07e60994579b64a6d992cf79210291b7d0b1b692f8f524516ed950872e5da10fb1b808b5a526dedc6fed1cf29807210386aa9372fbab374593466bc5451dc59954e90787f08060964d95c87ef34ca5bb53ae"
-	privkey, err := NewPrivkeyFromWif("cUfipPioYnHU61pfYTH9uuNoswRXx8rtzXhJZrsPeVV1LRFdTxvp")
+	privkey, err := key.NewPrivkeyFromWif("cUfipPioYnHU61pfYTH9uuNoswRXx8rtzXhJZrsPeVV1LRFdTxvp")
 	assert.NoError(t, err)
-	pubkey, err := privkey.GetPubkey()
+	privkeyUtil := &key.PrivkeyUtil{}
+	pubkey, err := privkeyUtil.GetPubkey(privkey)
 	assert.NoError(t, err)
 
 	// create pegin address
-	addrUtil := AddressUtil{}
-	peginAddr, claimScript, err := addrUtil.GetPeginAddressByPubkey(P2shP2wshAddress, fedpegScript, pubkey.Hex)
+	addrUtil := address.AddressUtil{}
+	peginAddr, claimScript, err := addrUtil.GetPeginAddressByPubkey(types.P2shP2wshAddress, fedpegScript, pubkey.Hex)
 	assert.NoError(t, err)
 	assert.Equal(t, "2MvmzAFKZ5xh44vyb7qY7NB2AoDuS55rVFW", peginAddr.Address)
 	assert.Equal(t, "0014e794713e386d83f32baa0e9d03e47c0839dc57a8", claimScript.ToHex())
@@ -252,17 +273,17 @@ func TestCreateClaimPeginTxByCfdConf(t *testing.T) {
 	feeAmount := int64(500)
 	peginAmount := amount - feeAmount
 	btcTxUtil := TransactionUtil{}
-	utxoOutPoint := OutPoint{
+	utxoOutPoint := types.OutPoint{
 		Txid: "ea9d5a9e974af1d167305aa6ee598706d63274e8a40f4f33af97db37a7adde4c",
 		Vout: 0,
 	}
-	btcInputs := []InputTxIn{
+	btcInputs := []types.InputTxIn{
 		{
 			OutPoint: utxoOutPoint,
-			Sequence: SequenceLockTimeFinal,
+			Sequence: types.SequenceLockTimeFinal,
 		},
 	}
-	btcOutputs := []InputTxOut{
+	btcOutputs := []types.InputTxOut{
 		{
 			Amount:  peginAmount,
 			Address: peginAddr.Address,
@@ -271,7 +292,7 @@ func TestCreateClaimPeginTxByCfdConf(t *testing.T) {
 	btcTx, err := btcTxUtil.Create(uint32(2), uint32(0), &btcInputs, &btcOutputs)
 	assert.NoError(t, err)
 	// add sign
-	utxos := []UtxoData{
+	utxos := []types.UtxoData{
 		{
 			Txid:       utxoOutPoint.Txid,
 			Vout:       utxoOutPoint.Vout,
@@ -279,34 +300,35 @@ func TestCreateClaimPeginTxByCfdConf(t *testing.T) {
 			Descriptor: "wpkh(02fd54c734e48c544c3c3ad1aab0607f896eb95e23e7058b174a580826a7940ad8)",
 		},
 	}
-	utxoPrivkey, err := NewPrivkeyFromWif("cNYKHjNc33ZyNMcDck59yWm1CYohgPhr2DYyCtmWNkL6sqb5L1rH")
+	utxoPrivkey, err := key.NewPrivkeyFromWif("cNYKHjNc33ZyNMcDck59yWm1CYohgPhr2DYyCtmWNkL6sqb5L1rH")
 	assert.NoError(t, err)
-	err = btcTx.SignWithPrivkey(&utxoOutPoint, utxoPrivkey, SigHashTypeAll, &utxos)
+	err = btcTxUtil.SignWithPrivkey(btcTx, &utxoOutPoint, utxoPrivkey, types.SigHashTypeAll, &utxos)
 	assert.NoError(t, err)
 	assert.Equal(t, "020000000001014cdeada737db97af334f0fa4e87432d6068759eea65a3067d1f14a979e5a9dea0000000000ffffffff010cdff5050000000017a91426b9ba9cf5d822b70cf490ad0394566f9db20c63870247304402200b3ca71e82551a333fe5c8ce9a8f8454eb8f08aa194180e5a87c79ccf2e46212022065c1f2a363ebcb155a80e234258394140d08f6ab807581953bb21a58f2d229a6012102fd54c734e48c544c3c3ad1aab0607f896eb95e23e7058b174a580826a7940ad800000000", btcTx.Hex)
 
-	assert.Equal(t, "12708508f0baf8691a3d7e22fd19afbf9bd8bf0d358e3310838bcc7916539c7b", btcTx.GetTxid())
+	assert.Equal(t, "12708508f0baf8691a3d7e22fd19afbf9bd8bf0d358e3310838bcc7916539c7b",
+		btcTxUtil.GetTxid(btcTx))
 
 	peginIndex := uint32(0)
 	txoutProof := "00000020fe3b574c1ce6d5cb68fc518e86f7976e599fafc0a2e5754aace7ca16d97a7c78ef9325b8d4f0a4921e060fc5e71435f46a18fa339688142cd4b028c8488c9f8dd1495b5dffff7f200200000002000000024a180a6822abffc3b1080c49016899c6dac25083936df14af12f58db11958ef27926299350fdc2f4d0da1d4f0fbbd3789d29f9dc016358ae42463c0cebf393f30105"
 
 	// create pegin tx
-	txUtil := ElementsTransactionUtil{}
-	peginOutPoint := OutPoint{
-		Txid: btcTx.GetTxid(),
+	txUtil := ConfidentialTxUtil{}
+	peginOutPoint := types.OutPoint{
+		Txid: btcTxUtil.GetTxid(btcTx),
 		Vout: peginIndex,
 	}
-	inputs := []InputConfidentialTxIn{
+	inputs := []types.InputConfidentialTxIn{
 		{
 			OutPoint: peginOutPoint,
-			PeginInput: &InputPeginData{
+			PeginInput: &types.InputPeginData{
 				BitcoinTransaction: btcTx.Hex,
 				ClaimScript:        claimScript.ToHex(),
 				TxOutProof:         txoutProof,
 			},
 		},
 	}
-	outputs := []InputConfidentialTxOut{
+	outputs := []types.InputConfidentialTxOut{
 		{
 			Amount:  99998500,
 			Address: "el1qqtl9a3n6878ex25u0wv8u5qlzpfkycc0cftk65t52pkauk55jqka0fajk8d80lafn4t9kqxe77cu9ez2dyr6sq54lwy009uex",
@@ -321,23 +343,23 @@ func TestCreateClaimPeginTxByCfdConf(t *testing.T) {
 	assert.Equal(t, "0200000001017b9c531679cc8b8310338e350dbfd89bbfaf19fd227e3d1a69f8baf0088570120000004000ffffffff020125b251070e29ca19043cf33ccd7324e2ddab03ecc4ae0b5e77c4fc0e5cf6c95a010000000005f5db2402fe5ec67a3f8f932a9c7b987e501f105362630fc2576d5174506dde5a94902dd7160014a7b2b1da77ffa99d565b00d9f7b1c2e44a6907a80125b251070e29ca19043cf33ccd7324e2ddab03ecc4ae0b5e77c4fc0e5cf6c95a0100000000000003e800000000000000000006080cdff505000000002025b251070e29ca19043cf33ccd7324e2ddab03ecc4ae0b5e77c4fc0e5cf6c95a2006226e46111a0b59caaf126043eb5bbf28c34f3a5e332a1fc7b2b73cf188910f160014e794713e386d83f32baa0e9d03e47c0839dc57a8c0020000000001014cdeada737db97af334f0fa4e87432d6068759eea65a3067d1f14a979e5a9dea0000000000ffffffff010cdff5050000000017a91426b9ba9cf5d822b70cf490ad0394566f9db20c63870247304402200b3ca71e82551a333fe5c8ce9a8f8454eb8f08aa194180e5a87c79ccf2e46212022065c1f2a363ebcb155a80e234258394140d08f6ab807581953bb21a58f2d229a6012102fd54c734e48c544c3c3ad1aab0607f896eb95e23e7058b174a580826a7940ad8000000009700000020fe3b574c1ce6d5cb68fc518e86f7976e599fafc0a2e5754aace7ca16d97a7c78ef9325b8d4f0a4921e060fc5e71435f46a18fa339688142cd4b028c8488c9f8dd1495b5dffff7f200200000002000000024a180a6822abffc3b1080c49016899c6dac25083936df14af12f58db11958ef27926299350fdc2f4d0da1d4f0fbbd3789d29f9dc016358ae42463c0cebf393f3010500000000", tx.Hex)
 
 	// blind
-	blindTxInList := []BlindInputData{
+	blindTxInList := []types.BlindInputData{
 		{
 			OutPoint: peginOutPoint,
 			Amount:   peginAmount,
 		},
 	}
-	option := NewBlindTxOption()
+	option := types.NewBlindTxOption()
 	option.AppendDummyOutput = true
-	err = tx.Blind(blindTxInList, nil, &option)
+	err = txUtil.Blind(tx, blindTxInList, nil, &option)
 	assert.NoError(t, err)
-	_, inList, outList, err := tx.GetAll(false)
+	_, inList, outList, err := txUtil.GetAll(tx, false)
 	assert.NoError(t, err)
 	assert.Equal(t, 1, len(inList))
 	assert.Equal(t, 3, len(outList))
 
 	// create signature
-	peginUtxos := []UtxoData{
+	peginUtxos := []types.UtxoData{
 		{
 			Txid:       peginOutPoint.Txid,
 			Vout:       peginOutPoint.Vout,
@@ -345,18 +367,18 @@ func TestCreateClaimPeginTxByCfdConf(t *testing.T) {
 			Descriptor: "wpkh(" + pubkey.Hex + ")",
 		},
 	}
-	sighash, err := tx.GetSighash(&peginOutPoint, SigHashTypeAll, &peginUtxos)
+	sighash, err := txUtil.GetSighash(tx, &peginOutPoint, types.SigHashTypeAll, &peginUtxos)
 	assert.NoError(t, err)
-	descriptor := &Descriptor{OutputDescriptor: peginUtxos[0].Descriptor}
-	signature, err := privkey.CreateEcSignature(sighash, &SigHashTypeAll)
+	desc := &types.Descriptor{OutputDescriptor: peginUtxos[0].Descriptor}
+	signature, err := privkeyUtil.CreateEcSignature(privkey, sighash, &types.SigHashTypeAll)
 	assert.NoError(t, err)
 
 	// add sign
-	err = tx.AddPubkeySignByDescriptor(&peginOutPoint, descriptor, signature.ToHex())
+	err = txUtil.AddPubkeySignByDescriptor(tx, &peginOutPoint, desc, signature.ToHex())
 	assert.NoError(t, err)
 
 	// verify
-	isVerify, reason, err := tx.VerifySign(&peginOutPoint, peginAmount, &peginUtxos)
+	isVerify, reason, err := txUtil.VerifySign(tx, &peginOutPoint, peginAmount, &peginUtxos)
 	assert.NoError(t, err)
 	assert.True(t, isVerify)
 	assert.Equal(t, "", reason)
@@ -365,55 +387,55 @@ func TestCreateClaimPeginTxByCfdConf(t *testing.T) {
 }
 
 func TestCreatePegoutTxByCfdConf(t *testing.T) {
-	SetCfdConfig(CfdConfig{
-		Network:                 NewNetworkTypeByString("liquidv1"),
+	config.SetCfdConfig(config.CfdConfig{
+		Network:                 types.NewNetworkTypeByString("liquidv1"),
 		BitcoinGenesisBlockHash: "0f9188f13cb7b2c71f2a335e3a4fc328bf5beb436012afca590b1a11466e2206",
 		BitcoinAssetId:          "5ac9f65c0efcc4775e0baec4ec03abdde22473cd3cf33c0419ca290e0751b225",
 	})
 
 	// mainchain address descriptor
 	mainchainXpubkey := "xpub6A53gzNdDBQYCtFFpZT7kUpoBGpzWigaukrdF9xoUZt7cYMD2qCAHVLsstNoQEDMFJWdX78KvT6yxpC76aGCN5mENVdWtFGcWZoKdtLq5jW"
-	mainchainPubkey, err := CfdGoGetPubkeyFromExtkey(mainchainXpubkey, int(KCfdNetworkMainnet))
+	mainchainPubkey, err := cfd.CfdGoGetPubkeyFromExtkey(mainchainXpubkey, int(cfd.KCfdNetworkMainnet))
 	assert.NoError(t, err)
-	negateMainchainPubkey, err := CfdGoNegatePubkey(mainchainPubkey)
+	negateMainchainPubkey, err := cfd.CfdGoNegatePubkey(mainchainPubkey)
 	assert.NoError(t, err)
 	mainchainOutputDescriptor := "pkh(" + mainchainXpubkey + "/0/*)"
 	bip32Counter := uint32(0)
 
 	onlinePrivkey := "L52AgshDAE14NHJuovwAw8hyrTNK4YQjuiPC9EES4sfM7oBPzU4o"
-	onlinePubkey, err := CfdGoGetPubkeyFromPrivkey("", onlinePrivkey, true)
+	onlinePubkey, err := cfd.CfdGoGetPubkeyFromPrivkey("", onlinePrivkey, true)
 	assert.NoError(t, err)
 	// whitelist
 	pakEntry := negateMainchainPubkey + onlinePubkey
 	whitelist := pakEntry
 
 	// pegout address
-	addrUtil := AddressUtil{}
-	pegoutAddr, baseDescriptor, err := addrUtil.GetPegoutAddress(P2pkhAddress, mainchainOutputDescriptor, bip32Counter)
+	addrUtil := address.AddressUtil{}
+	pegoutAddr, baseDescriptor, err := addrUtil.GetPegoutAddress(types.P2pkhAddress, mainchainOutputDescriptor, bip32Counter)
 	assert.NoError(t, err)
 	assert.Equal(t, "1NrcpiZmCxjC7KVKAYT22SzVhhcXtp5o4v", pegoutAddr.Address)
 	assert.Equal(t, "pkh("+mainchainXpubkey+")", *baseDescriptor)
 
 	// create pegout tx
-	txUtil := ElementsTransactionUtil{}
+	txUtil := ConfidentialTxUtil{}
 	pegoutAddrList := []string{}
-	inputs := []InputConfidentialTxIn{
+	inputs := []types.InputConfidentialTxIn{
 		{
-			OutPoint: OutPoint{
+			OutPoint: types.OutPoint{
 				Txid: "4aa201f333e80b8f62ba5b593edb47b4730212e2917b21279f389ba1c14588a3",
 				Vout: 0,
 			},
 			Sequence: 4294967293,
 		},
 	}
-	outputs := []InputConfidentialTxOut{
+	outputs := []types.InputConfidentialTxOut{
 		{
 			Amount:  209998999992700,
 			Address: "XBMr6srTXmWuHifFd8gs54xYfiCBsvrksA",
 		},
 		{
 			Amount: 1000000000,
-			PegoutInput: &InputPegoutData{
+			PegoutInput: &types.InputPegoutData{
 				OnlineKey:               onlinePrivkey,
 				BitcoinOutputDescriptor: mainchainOutputDescriptor,
 				Bip32Counter:            bip32Counter,
@@ -431,7 +453,7 @@ func TestCreatePegoutTxByCfdConf(t *testing.T) {
 	assert.Equal(t, 1, len(pegoutAddrList))
 	assert.Equal(t, "1NrcpiZmCxjC7KVKAYT22SzVhhcXtp5o4v", pegoutAddrList[0])
 
-	pegoutAddress, hasPegout, err := tx.GetPegoutAddress(uint32(1))
+	pegoutAddress, hasPegout, err := txUtil.GetPegoutAddress(tx, uint32(1))
 	assert.NoError(t, err)
 	assert.True(t, hasPegout)
 	assert.Equal(t, "1NrcpiZmCxjC7KVKAYT22SzVhhcXtp5o4v", pegoutAddress.Address)
