+++ config/config.go
@@ -1,10 +1,15 @@
-package cfdgo
+package config
 
-import "fmt"
+import (
+	"fmt"
+
+	cfd "github.com/cryptogarageinc/cfd-go"
+	types "github.com/cryptogarageinc/cfd-go/types"
+)
 
 // GlobalConfig This struct is cfd's global configuration.
 type CfdConfig struct {
-	Network                 NetworkType
+	Network                 types.NetworkType
 	BitcoinGenesisBlockHash string
 	BitcoinAssetId          string
 }
@@ -28,7 +33,7 @@ func SetCfdConfig(config CfdConfig) error {
 		return fmt.Errorf("CFD Error: Invalid network type")
 	}
 	if len(config.BitcoinGenesisBlockHash) > 0 {
-		if _, err := NewByteDataFromHex(config.BitcoinGenesisBlockHash); err != nil {
+		if _, err := cfd.NewByteDataFromHex(config.BitcoinGenesisBlockHash); err != nil {
 			return err
 		} else if len(config.BitcoinGenesisBlockHash) != 64 {
 			return fmt.Errorf("CFD Error: Invalid blockHash length")
@@ -36,7 +41,7 @@ func SetCfdConfig(config CfdConfig) error {
 		cfdConfig.BitcoinGenesisBlockHash = config.BitcoinGenesisBlockHash
 	}
 	if len(config.BitcoinAssetId) > 0 {
-		if _, err := NewByteDataFromHex(config.BitcoinAssetId); err != nil {
+		if _, err := cfd.NewByteDataFromHex(config.BitcoinAssetId); err != nil {
 			return err
 		} else if len(config.BitcoinAssetId) != 64 {
 			return fmt.Errorf("CFD Error: Invalid assetId length")
