+++ go_test.bat
@@ -1,4 +1,4 @@
 @echo off
 setlocal
 set PATH=%PATH%;%~dp0%\build\Release;
-call go test
+call go test . ./apis/transaction ./apis/block ./tests -v
