+++ types/address.go
@@ -1,0 +1,8 @@
+package types
+
+// Address ...
+type Address struct {
+	Address string
+	Network NetworkType
+	Type    AddressType
+}
