+++ types/block.go
@@ -1,0 +1,6 @@
+package types
+
+// Block ...
+type Block struct {
+	Hex string
+}
