+++ types/confidential_address.go
@@ -1,0 +1,10 @@
+package types
+
+// ConfidentialAddress ...
+type ConfidentialAddress struct {
+	ConfidentialAddress string
+	Address             string
+	Network             NetworkType
+	Type                AddressType
+	ConfidentialKey     *Pubkey
+}
