+++ types/privkey.go
@@ -1,0 +1,9 @@
+package types
+
+// Privkey ...
+type Privkey struct {
+	Hex                string
+	Wif                string
+	Network            NetworkType
+	IsCompressedPubkey bool
+}
