+++ types/pubkey.go
@@ -1,0 +1,6 @@
+package types
+
+// Pubkey ...
+type Pubkey struct {
+	Hex string
+}
